+++ packages/backend/test/core/tx-count/RpcTransactionUpdater.test.ts
@@ -6,21 +6,25 @@ import { BigNumber } from 'ethers'
 import waitForExpect from 'wait-for-expect'
 
 import { Clock } from '../../../src/core/Clock'
-import { BlockTxCountUpdater } from '../../../src/core/tx-count/BlockTxCountUpdater'
-import { TxCountRepository } from '../../../src/peripherals/database/TxCountRepository'
+import { RpcTransactionUpdater } from '../../../src/core/tx-count/RpcTransactionUpdater'
+import { TransactionCountRepository } from '../../../src/peripherals/database/TransactionCountRepository'
 import { EthereumClient } from '../../../src/peripherals/ethereum/EthereumClient'
-import { fakeTxCount } from '../../peripherals/database/TxCountRepository.test'
+import { fakeTxCount } from '../../peripherals/database/TransactionCountRepository.test'
 
-describe(BlockTxCountUpdater.name, () => {
-  describe(BlockTxCountUpdater.prototype.start.name, () => {
+describe(RpcTransactionUpdater.name, () => {
+  describe(RpcTransactionUpdater.prototype.start.name, () => {
     it('skips known blocks', async () => {
       const ethereumClient = mock<EthereumClient>({
         getBlock: async () => fakeBlock(),
         getBlockNumber: async () => 5n,
       })
-      const txCountRepository = mock<TxCountRepository>({
+      const txCountRepository = mock<TransactionCountRepository>({
         findLatestByProject: async () => fakeTxCount({ blockNumber: 4 }),
-        getMissingRangesByProject: async () => [[1, 3]],
+        getMissingRangesByProject: async () => [
+          [-Infinity, -1],
+          [2, 3],
+          [5, Infinity],
+        ],
         add: async () => '',
       })
       const clock = mock<Clock>({
@@ -30,7 +34,7 @@ describe(BlockTxCountUpdater.name, () => {
         },
         getLastHour: () => UnixTime.now(),
       })
-      const blockTxCountUpdater = new BlockTxCountUpdater(
+      const blockTxCountUpdater = new RpcTransactionUpdater(
         ethereumClient,
         txCountRepository,
         clock,
@@ -45,15 +49,19 @@ describe(BlockTxCountUpdater.name, () => {
     })
   })
 
-  describe(BlockTxCountUpdater.prototype.update.name, () => {
+  describe(RpcTransactionUpdater.prototype.update.name, () => {
     it('does not query the same blocks multiple times', async () => {
       const ethereumClient = mock<EthereumClient>({
         getBlock: async () => fakeBlock(),
         getBlockNumber: async () => 5n,
       })
-      const txCountRepository = mock<TxCountRepository>({
+      const txCountRepository = mock<TransactionCountRepository>({
         findLatestByProject: async () => fakeTxCount({ blockNumber: 4 }),
-        getMissingRangesByProject: async () => [[1, 3]],
+        getMissingRangesByProject: async () => [
+          [-Infinity, -1],
+          [2, 3],
+          [5, Infinity],
+        ],
         add: async () => '',
       })
       const clock = mock<Clock>({
@@ -63,7 +71,7 @@ describe(BlockTxCountUpdater.name, () => {
         },
         getLastHour: () => UnixTime.now(),
       })
-      const blockTxCountUpdater = new BlockTxCountUpdater(
+      const blockTxCountUpdater = new RpcTransactionUpdater(
         ethereumClient,
         txCountRepository,
         clock,
@@ -79,7 +87,7 @@ describe(BlockTxCountUpdater.name, () => {
     })
   })
 
-  describe(BlockTxCountUpdater.prototype.getBlock.name, () => {
+  describe(RpcTransactionUpdater.prototype.updateBlock.name, () => {
     it('downloads and saves a block to DB', async () => {
       const TIME_0 = UnixTime.now().add(-1, 'days').toStartOf('day')
       const TIME_1 = TIME_0.add(1, 'hours')
@@ -101,13 +109,13 @@ describe(BlockTxCountUpdater.name, () => {
           ),
         getBlockNumber: async () => 5n,
       })
-      const txCountRepository = mock<TxCountRepository>({
+      const txCountRepository = mock<TransactionCountRepository>({
         add: async () => '',
       })
       const clock = mock<Clock>({
         getLastHour: () => UnixTime.now(),
       })
-      const blockTxCountUpdater = new BlockTxCountUpdater(
+      const blockTxCountUpdater = new RpcTransactionUpdater(
         ethereumClient,
         txCountRepository,
         clock,
@@ -115,8 +123,8 @@ describe(BlockTxCountUpdater.name, () => {
         ProjectId('fake-project'),
       )
 
-      await blockTxCountUpdater.getBlock(1)
-      await blockTxCountUpdater.getBlock(2)
+      await blockTxCountUpdater.updateBlock(1)
+      await blockTxCountUpdater.updateBlock(2)
 
       expect(txCountRepository.add).toHaveBeenCalledExactlyWith([
         [
